feat: add input redirections
