+++ Makefile
@@ -1,7 +1,7 @@
 NAME  = minishell
 VPATH = ./src \
 		./src/parser ./src/prompt ./src/signals \
- 		./src/here_doc ./src/utils \
+ 		./src/utils \
  		./src/handler_quotes \
  		./src/tokenizer ./src/expansions \
  		./src/syntax ./src/env ./src \
@@ -9,7 +9,7 @@ VPATH = ./src \
 
 CFLAGS = -g -Wall -Wextra -Werror
 SOURCES = main.c wait_input.c tokenizer.c setup_hook.c\
-		  heredoc.c check_unclosed.c display.c ft_linked_list.c \
+		  check_unclosed.c display.c ft_linked_list.c \
 	  	  init_signal.c clear_bimatrix.c retokenize.c handle_expansions.c \
 		  find_token.c strings_utils.c strjoinfree_s1.c print_data.c\
 		  tokenizer_operator.c init_parsers.c linked_list.c \
@@ -19,10 +19,10 @@ SOURCES = main.c wait_input.c tokenizer.c setup_hook.c\
 		  env_builtin.c exit_builtin.c executor.c export_builtin.c \
 		  unset_builtin.c utils_builtins_01.c echo_builtin.c utils_builtins_02.c \
 		  init_env.c init_statics.c is_llmin.c heredoc_utils.c pwd_builtin.c \
-		  expansion_utils.c print_tokens.c init_redirections.c \
+		  init_t_file_io.c init_redirections.c open_redirections.c \
+		  expansion_utils.c print_tokens.c heredoc.c\
 		  utils_executor.c one_command.c exec_command.c get_cmd_path.c \
 		  split_envp.c command_setuphook.c pre_executor.c is_a_command.c \
-		  init_t_file_io.c
 
 LIB_SRCS = ft_strlen.c ft_strdup.c ft_substr.c ft_itoa.c \
 				 ft_split.c ft_strlcpy.c ft_isalpha.c \
