+++ src/executor/pre_executor.c
@@ -6,7 +6,7 @@
 /*   By: johmatos <johmatos@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/07 20:10:54 by johmatos          #+#    #+#             */
-/*   Updated: 2023/07/07 20:11:01 by johmatos         ###   ########.fr       */
+/*   Updated: 2023/07/14 14:09:14 by johmatos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,10 +16,10 @@ static int	init_redir(void)
 {
 	int	status;
 
-	status = init_heredoc(getter_data()->cmds->head);
+	status = init_input(getter_data()->cmds->head);
 	if (WEXITSTATUS(status) == 129)
 		return (status);
-	status = init_redirections(getter_data()->cmds->head);
+	status = init_output(getter_data()->cmds->head);
 	return (status);
 }
