+++ src/redirections/init_redirections.c
@@ -6,47 +6,12 @@
 /*   By: johmatos <johmatos@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/20 19:46:33 by johmatos          #+#    #+#             */
-/*   Updated: 2023/07/12 20:16:49 by johmatos         ###   ########.fr       */
+/*   Updated: 2023/07/14 14:23:22 by johmatos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include "minishell.h"
-#include <stdio.h>
-
-static int	setup_out_redir(t_node *node, t_tokens token)
-{
-	int		fd;
-	char	*path;
-
-	while (node->token != T_WORD)
-		node = node->next;
-	path = node->str;
-	if (token == T_OUT_REDIR)
-		fd = open(path, O_RDWR | O_TRUNC | O_CREAT, 0666);
-	else
-		fd = open(path, O_RDWR | O_CREAT | O_APPEND, 0666);
-	return (fd);
-}
-
-static void	open_redir(t_node *node, t_process_io *fds, int *status)
-{
-	int	aux;
-
-	aux = 0;
-	while (node != NULL && *status != 2)
-	{
-		if (node->token == T_PIPE)
-			aux++;
-		if (node->token == T_OUT_REDIR || node->token == T_O_OUT_REDIR)
-		{
-			if (fds[aux].output)
-				close(fds[aux].output);
-			fds[aux].output = setup_out_redir(node, node->token);
-		}
-		node = node->next;
-	}
-}
 
 int	*getter_outputs(void)
 {
@@ -55,13 +20,33 @@ int	*getter_outputs(void)
 	return (redirections);
 }
 
-int	init_redirections(t_node *node)
+int	*getter_inputs(void)
+{
+	static int	heredoc_fds[MAX_FD];
+
+	ft_bzero(heredoc_fds, sizeof(*heredoc_fds));
+	return (heredoc_fds);
+}
+
+int	init_output(t_node *node)
 {
 	t_process_io	*redirections;
 	int				status;
 
 	status = 0;
 	redirections = getter_t_process_io();
-	open_redir(node, redirections, &status);
+	open_out_redir(node, redirections, &status);
+	return (status);
+}
+
+int	init_input(t_node *node)
+{
+	t_process_io	*heredoc_fd;
+	int				status;
+
+	heredoc_fd = getter_t_process_io();
+	status = 0;
+	ft_bzero(heredoc_fd, MAX_FD);
+	open_input_redir(node, heredoc_fd, &status);
 	return (status);
 }
